// Source/Mobile/src/app/models/address.model.ts
export class Address {
    uuid: string = '';
    uid: string = '';
    store_id: string = '';
    type: string = '';
    address: string = '';
    house: string = '';
    landmark: string = '';
    zipcode: string = '';
    lat: string = '';
    lng: string = '';
    status: string = '';
    updated_at: string = '';
    deleted_at: string = '';
}